DA files moved inside specific folder
